refactor: Score 객체 삭제 및 BlackjackGame에 합침
